# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for inovantapp.",
        "title": "inovantapp API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {},
        "version": "1.0"
    },
    "host": "192.168.1.69:8080",
    "basePath": "/",
    "paths": {
        "/api/actions-verification": {
            "get": {
                "description": "Get ActionVerification list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ActionVerification.List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Context to return",
                        "name": "context",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter ActionVerifications by type [acveID]",
                        "name": "acveID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter ActionVerifications by type [userID]",
                        "name": "userID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter ActionVerifications by type [type]",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter ActionVerifications by type [verification]",
                        "name": "verification",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter ActionVerifications by type [createdAt[gte]]",
                        "name": "createdAt[gte]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter ActionVerifications by type [createdAt[lte]]",
                        "name": "createdAt[lte]",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.actionsVerificationListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create ActionVerification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ActionVerification.Create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Context to return",
                        "name": "context",
                        "in": "query"
                    },
                    {
                        "description": "Create new ActionVerification",
                        "name": "ActionVerification",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.ActionVerification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.actionVerificationGetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/actions-verification/{acveID}": {
            "get": {
                "description": "Get a ActionVerification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ActionVerification.Get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Context to return",
                        "name": "context",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter ActionVerifications by type [acveID]",
                        "name": "acveID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.actionVerificationGetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update ActionVerification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ActionVerification.Update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Context to return",
                        "name": "context",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "string",
                        "description": "ActionVerification ID",
                        "name": "acveID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ActionVerification Update Body",
                        "name": "ActionVerification",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.ActionVerification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.actionVerificationGetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.actionVerificationGetResponse": {
            "type": "object",
            "properties": {
                "context": {
                    "description": "Client sets this value and server echos data in the response",
                    "type": "string"
                },
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/handler.actionVerificationResponse"
                }
            }
        },
        "handler.actionVerificationResponse": {
            "type": "object",
            "properties": {
                "item": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                }
            }
        },
        "handler.actionsVerificationListResponse": {
            "type": "object",
            "properties": {
                "context": {
                    "description": "Client sets this value and server echos data in the response",
                    "type": "string"
                },
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/handler.actionsVerificationResponse"
                }
            }
        },
        "handler.actionsVerificationResponse": {
            "type": "object",
            "properties": {
                "currentItemCount": {
                    "description": "The number of items in this result set",
                    "type": "integer",
                    "example": 1
                },
                "items": {
                    "type": "string"
                },
                "itemsPerPage": {
                    "description": "The number of items in the result",
                    "type": "integer",
                    "example": 10
                },
                "kind": {
                    "type": "string"
                },
                "lang": {
                    "description": "Indicates the language of the rest of the properties in this object (BCP 47)",
                    "type": "string",
                    "example": "pt-br"
                },
                "pageIndex": {
                    "description": "The index of the current page of items",
                    "type": "integer",
                    "example": 1
                },
                "startIndex": {
                    "description": "The index of the first item in data.items",
                    "type": "integer",
                    "example": 1
                },
                "totalItems": {
                    "description": "The total number of items available in this set",
                    "type": "integer",
                    "example": 100
                },
                "totalPages": {
                    "description": "The total number of pages in the result set.",
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "handler.detailError": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "extendedHelp": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "locationType": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "sendReport": {
                    "type": "string"
                }
            }
        },
...
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "$ref": "#/definitions/handler.generalError"
                }
            }
        },
...
        "handler.generalError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.detailError"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.ActionVerification": {
            "type": "object",
            "properties": {
                "acveID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                },
                "verification": {
                    "type": "string"
                }
            }
        },
...
    }
}